move functions to form genotype-order.js
add genotype-order.js, with :
  MatchRefSample and tsneOrder() from vcf-feature.js, and
  referenceSamplesSymbol, referenceSampleMatchesSymbol, distancesTo1d() from manage-genotype.js
The motive for moving these functions is that this is evolving into a focus of functionality and will expand, and manage-genotype.js and vcf-feature.js are already fairly large.
